feat: move addValidationTriplesToFormNodesL to an utils method
Refs: #SFB-8
